requirements.txt
Django==1.11
Pillow==3.1.2
numpy==1.12.1
scipy==0.19.0
scikit-image==0.13.0
gunicorn
psycopg2
whitenoise
django-cors-headers
matplotlib